Готова мобильная версия + берется информация из json server
