+++ src/MobileComponents/Mobile.module.scss
@@ -97,7 +97,7 @@ display: flex;
 .picAndName {
   display: flex;
   flex-direction: row;
-  gap: 15px;
+  gap: 10px;
 }
 .linePic {
   display: flex;
