+++ src/content2/topCollection/topCollection.module.scss
@@ -18,6 +18,8 @@
     text-align: center;
     color: black;
     font-size: 45px;
+    margin-bottom: 80px;
+    font-weight: 500;
   }
 
   .title {
