+++ src/content3/exploreMarketplace/exploreMarketplace.module.scss
@@ -49,7 +49,7 @@ $max-width: 375px;
 .line {
   display: grid;
   grid-template-columns: repeat(4, auto);
-  gap: 4vw;
+  gap: 3vw;
 }
 @media screen and (max-width: #{$max-width}) {
   .line {
